Now saves lot after verifying the information instead of before.
